requirements.txt

---

### `requirements.txt`
```txt
torch>=2.1
torchvision>=0.16
torchaudio>=2.1
transformers>=4.42
sentencepiece
datasets>=2.20
scikit-learn>=1.3
numpy>=1.24
pandas>=2.0
librosa>=0.10
soundfile>=0.12
opencv-python>=4.9
facenet-pytorch>=2.5
mediapipe>=0.10
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-multipart>=0.0.9
faiss-cpu>=1.8
rank-bm25>=0.2
nltk>=3.8
streamlit>=1.33
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.8
